[products] CRUD de produtos e validação de exclusão de categorias e eventos.

+++ src/features/category/application/use-cases/category-remove.use-case.ts
@@ -1,9 +1,11 @@
 import { Application } from '@/common/application/use-cases/application';
 import { AbilitiesEnum } from '@/common/infra/enums/abilities.enum';
 import { CategoryValidations } from '@/features/category/application/validations/category.validations';
-import { Inject, Injectable } from '@nestjs/common';
+import { BadRequestException, Inject, Injectable } from '@nestjs/common';
 import { CategoryRepository } from '@/features/category/domain/repositories/category.repository';
 import { AbstractCategoryRemoveUseCase } from '@/features/category/domain/use-cases/abstract.category-remove.use-case';
+import { ErrorMessagesEnum } from '@/common/infra/enums/error-messages.enum';
+import { Category } from '@/features/category/domain/core/category';
 
 @Injectable()
 export class CategoryRemoveUseCase
@@ -20,10 +22,16 @@ export class CategoryRemoveUseCase
   async execute(uuid: string): Promise<void> {
     this.policy.can(AbilitiesEnum.CATEGORIES_DELETE);
 
-    const category = await CategoryValidations.categoryExists(
+    const category: Category = await CategoryValidations.categoryExists(
       uuid,
       this.categoryRepository,
     );
+
+    if (category.products.length > 0) {
+      throw new BadRequestException(
+        ErrorMessagesEnum.CATEGORY_HAS_PRODUCTS_IN_DELETE,
+      );
+    }
 
     await this.categoryRepository.remove(category.uuid);
   }
